requirements.txt
mcp[cli]>=1.14.1
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
httpx>=0.24.0
python-dateutil>=2.8.0
